Add type casting from string to Path in results.py
Previously, an error was thrown in som cases when a pathlib.Path is expected by a string is supplied, for example if  in . Now, if the supplied path is not of type pathlib.Path, we type cast it to the expected form. This will raise an error in case it is not possible to type cast.